Revert "fix: resolve compilation error through restructuring"
