+++ lib/batched_file_sink_impl.h
@@ -9,11 +9,7 @@
 #define INCLUDED_SPECTRE_BATCHED_FILE_SINK_IMPL_H
 
 #include <gnuradio/spectre/batched_file_sink.h>
-
-#include <fstream>
-#include <filesystem>
-#include <chrono>
-#include <ctime>
+#include <gnuradio/spectre/bin_chunk_helpers.h>
 
 namespace fs = std::filesystem;
 
@@ -52,27 +48,6 @@ public:
         gr_vector_const_void_star& input_items,
         gr_vector_void_star& output_ite
     );
-};
-
-class bin_chunk_helper
-{
-private:
-    fs::path _parent_path; /**< Defines the absolute path to the parent directory to hold the batched files. */
-    std::string _tag; /**< File names are tagged with this string. */
-    int32_t _millisecond_correction; /**< A 32-bit integer describing the millisecond correction for the timestamp of the first sample in each batch.*/
-    fs::path _absolute_path; /**< The absolute path to the binary chunk file. */
-
-public:
-    bin_chunk_helper(
-        std::string parent_path,
-        std::string tag
-    );
-
-    ~bin_chunk_helper();
-    void set_attrs(); /**< Effectively a setter for the _millisecond_correction and _absolute_path member variables. */
-    fs::path get_absolute_path(); /**< Public getter for the _absolute_path member variable. */
-    int32_t get_millisecond_correction(); /**< Public getter for the _millisecond_correction member variable. */
-    fs::path get_absolute_parent_path(); /**< Public getter for the parent path of the binary chunk file.*/
 };
 
 } // namespace spectre
